fix: 故障扣分类型

+++ dal/src/main/java/com/timevale/forward/dal/entity/TroubleTicketDO.java
@@ -167,7 +167,7 @@ public class TroubleTicketDO extends BaseDO {
     /**
      * 故障扣分
      */
-    private Integer deductPoints;
+    private Long deductPoints;
 
     /**
      * 数据统计说明
